WEB-4460 move modals into directory
