+++ app/pages/clinicworkspace/TideDashboardV2/TideDashboardV2.js
@@ -21,8 +21,8 @@ import EmptyContentNode from './EmptyContentNode';
 import { Redirect, useLocation } from 'react-router-dom';
 import useAuthorizationGate from './useAuthorizationGate';
 
-import EditPatientDialogController from './EditPatientDialogController';
-import DataConnectionsModalController from './DataConnectionsModalController';
+import EditPatientDialogController from './modals/EditPatientDialogController';
+import DataConnectionsModalController from './modals/DataConnectionsModalController';
 
 const Gap = () => <Box sx={{ marginLeft: 'auto' }}></Box>;
